Extracted 3 dialogs to their own components
